+++ src/rajawali/animation/ScaleAnimation3D.java
@@ -1,3 +1,15 @@
+/**
+ * Copyright 2013 Dennis Ippel
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ * 
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
 package rajawali.animation;
 
 import rajawali.ATransformable3D;
@@ -12,11 +24,11 @@ public class ScaleAnimation3D extends Animation3D {
 	protected Vector3 mAddedScale = new Vector3();
 
 	
-	public ScaleAnimation3D(float toScale) {
+	public ScaleAnimation3D(double toScale) {
 		super();
 		mToScale = new Vector3(toScale);
 	}
-	public ScaleAnimation3D(float fromScale, float toScale) {
+	public ScaleAnimation3D(double fromScale, double toScale) {
 		super();
 		mToScale = new Vector3(toScale);
 		mFromScale = new Vector3(fromScale);
@@ -45,7 +57,7 @@ public class ScaleAnimation3D extends Animation3D {
 		if (mDiffScale == null)
 			mDiffScale = Vector3.subtractAndCreate(mToScale, mFromScale);
 
-		mMultipliedScale.scaleAndSet(mDiffScale, (float) mInterpolatedTime);
+		mMultipliedScale.scaleAndSet(mDiffScale, mInterpolatedTime);
 		mAddedScale.addAndSet(mFromScale, mMultipliedScale);
 		mTransformable3D.setScale(mAddedScale);
 	}
