+++ src/rajawali/animation/SlerpAnimation3D.java
@@ -1,3 +1,15 @@
+/**
+ * Copyright 2013 Dennis Ippel
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ * 
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
 package rajawali.animation;
 
 import rajawali.math.MathUtil;
@@ -29,11 +41,10 @@ public class SlerpAnimation3D extends Animation3D {
 	private Vector3 mTmpVec;
 	private Vector3 mTmpQuatVector;
 	private Quaternion mTmpQuat;
-	private float[] mRotationMatrix;
-	private float mDistance;
+	private double[] mRotationMatrix;
+	private double mDistance;
 	
-	public SlerpAnimation3D(Vector3 from, Vector3 to)
-	{
+	public SlerpAnimation3D(Vector3 from, Vector3 to) {
 		super();
 		mFrom = quaternionFromVector(from.clone());
 		mTo = quaternionFromVector(to.clone());
@@ -41,12 +52,12 @@ public class SlerpAnimation3D extends Animation3D {
 		mTmpQuatVector = new Vector3();
 		mTmpQuat = new Quaternion();
 		mDistance = from.length();
-		mRotationMatrix = new float[16];
+		mRotationMatrix = new double[16];
 	}
 	
 	@Override
 	protected void applyTransformation() {
-		mTmpQuat.slerp(mFrom, mTo, (float)mInterpolatedTime);
+		mTmpQuat.slerp(mFrom, mTo, mInterpolatedTime);
 		mTmpVec.setAll(mForwardVec);
 		mTmpQuat.toRotationMatrix(mRotationMatrix);
 		mTmpVec.multiply(mRotationMatrix);
@@ -54,10 +65,9 @@ public class SlerpAnimation3D extends Animation3D {
 		mTransformable3D.setPosition(mTmpVec);
 	}
 	
-	private Quaternion quaternionFromVector(Vector3 vec)
-	{
+	private Quaternion quaternionFromVector(Vector3 vec) {
 		vec.normalize();
-		float angle = MathUtil.radiansToDegrees((float)Math.acos(Vector3.dot(mForwardVec, vec)));
+		double angle = MathUtil.radiansToDegrees(Math.acos(Vector3.dot(mForwardVec, vec)));
 		Quaternion q = new Quaternion();
 		q.fromAngleAxis(mTmpQuatVector.crossAndSet(mForwardVec, vec), angle);
 		return q;
