+++ src/rajawali/materials/PhongMaterial.java
@@ -1,3 +1,15 @@
+/**
+ * Copyright 2013 Dennis Ippel
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ * 
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
 package rajawali.materials;
 
 import rajawali.lights.ALight;
@@ -53,9 +65,9 @@ public class PhongMaterial extends AAdvancedMaterial {
 	}
 
 	public void setSpecularColor(Vector3 color) {
-		mSpecularColor[0] = color.x;
-		mSpecularColor[1] = color.y;
-		mSpecularColor[2] = color.z;
+		mSpecularColor[0] = (float) color.x;
+		mSpecularColor[1] = (float) color.y;
+		mSpecularColor[2] = (float) color.z;
 		mSpecularColor[3] = 1;
 	}
